Module Rope

A rope is a standard data structure that represents a single string as a tree of strings, allowing concatenation to do no work up front. See the README.md file for details and motivating examples.

type t

include Ppx_compare_lib.Equal.S with type t := t

val equal : t -> t -> bool

include Ppx_quickcheck_runtime.Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val of_string : Base.string -> t

Takes O(1) time. The string isn't copied, so don't mutate it.

val empty : t

val is_empty : t -> Base.bool

val length : t -> Base.int

val to_string : t -> Base.string

Allocates a fresh string, so takes time proportional to the total size of the result.

val to_char_sequence : t -> Base.char Base.Sequence.t

to_char_sequence can often produce characters incrementally, but in the worst case it takes time and memory proportional to the total length of the string to produce even a single character. (In such cases, it should still only take O(length) time to produce the rest of the string.)

val (^) : t -> t -> t

These take time proportional to the number of t's passed.

val concat : ?sep:t -> t Base.list -> t

val concat_array : ?sep:t -> t Base.array -> t

val add_to_buffer : t -> Base.Buffer.t -> Base.unit

Appends the contents of the Rope at the end of a destination buffer.

val is_prefix : t -> prefix:t -> Base.bool

is_prefix a ~prefix:b is a more efficient version of String.is_prefix (Rope.to_string a) ~prefix:(Rope.to_string b). However, the worst-case time complexity is still O(length t) instead of O(min(length t, length prefix)) as one could expect.

module For_testing : sig ... end